+++ Tests/MapboxSearchTests/Common/Models/Metadata/OpenHoursTests.swift
@@ -27,7 +27,7 @@ class OpenHoursTests: XCTestCase {
         let coreHours = CoreOpenHours(mode: .scheduled, periods: .coreOpenHourPeriods(), weekdayText: nil, note: nil)
         let hours = OpenHours(coreHours)
 
-        if case .scheduled(let periods) = hours {
+        if case .scheduled(let periods, _, _) = hours {
             XCTAssertEqual(periods.count, 7)
         } else {
             XCTFail("Expected `.scheduled` case")
@@ -44,16 +44,53 @@ class OpenHoursTests: XCTestCase {
         XCTAssertEqual(hours, .permanentlyClosed)
     }
 
-    func testScheduledCaseCodableSupport() throws {
+    /// Test Scheduled Case returned from SBS API
+    func testScheduledCaseSBSCodableSupport() throws {
         let scheduled = OpenHours.scheduled(periods: .fullWeekPeriodsSample())
 
         let data = try jsonEncoder.encode(scheduled)
         let decodedScheduled = try jsonDecoder.decode(OpenHours.self, from: data)
 
-        if case .scheduled(let periods) = decodedScheduled {
+        XCTAssertEqual(decodedScheduled, scheduled)
+        if case .scheduled(let periods, _, _) = decodedScheduled {
             XCTAssertEqual(periods, .fullWeekPeriodsSample())
         } else {
             XCTFail("Expected `.scheduled` case")
         }
+    }
+
+    /// Test Scheduled Case returned from SearchBox API
+    func testScheduledCaseSearchBoxCodableSupport() throws {
+        let scheduled = OpenHours.scheduled(
+            periods: .fullWeekPeriodsSample(),
+            weekdayText: .fullWeekdayTextSample(),
+            note: "Value present"
+        )
+
+        let data = try jsonEncoder.encode(scheduled)
+        let decodedScheduled = try jsonDecoder.decode(OpenHours.self, from: data)
+
+        XCTAssertEqual(decodedScheduled, scheduled)
+        if case .scheduled(let periods, let weekdayText, let note) = decodedScheduled {
+            XCTAssertEqual(periods, .fullWeekPeriodsSample())
+            XCTAssertEqual(weekdayText, .fullWeekdayTextSample())
+            XCTAssertEqual("Value present", note)
+        } else {
+            XCTFail("Expected `.scheduled` case")
+        }
+    }
+}
+
+extension OpenHours.WeekdayText {
+    public static func fullWeekdayTextSample() -> Self {
+        [
+            "Monday 08:35am - 8:18pm",
+            "Tuesday 08:35am - 8:18pm",
+            "Wednesday 08:35am - 8:18pm",
+            "Thursday 08:35am - 8:18pm",
+            "Friday 08:35am - 8:18pm",
+            "Saturday 10:00am - 7:00pm",
+            "Sunday 10:00am - 7:00pm",
+        ]
     }
 }